OOP approach
now there is one file for the chapter object and another for the
"page translator". they still are deeply connected as we pass around
both the reader and the deepl client to avoid many api calls.

this meant also splitting code in some files but the result is that we
just import a function into "main.py" and go with it.

we could just merge "interfaces.py, translator.py, chapter.py" in a
single one.

before doing anything: TEST IT !