remove tqdm and generators
tqdm is really not needed and for a simple progress bar we can do that
on our own.

removing the generators in favor of passing list around all the time
is easier for both type hints and data management (and also progress
bar creation, as we always know how many elements we are dealing with).
Yes, we are using some more memory this way but this is Python and this
kind of optimization should not be a concern of ours, at least at this
stage of development.

Only "alternative.py" was modified, the rest of the code is untouched.
I did it directly on the main branch just so LS can see the changes
immediately.